Splitted our functional test into multiple files
