+++ geckodriver.log
@@ -4717,3 +4717,45 @@ See /usr/include/servers/bootstrap_defs.h for the error codes.
 1511381696017	Marionette	DEBUG	Received DOM event "DOMContentLoaded" for "http://localhost:55197/accounts/register/"
 1511381696143	Marionette	DEBUG	Received DOM event "pageshow" for "http://localhost:55197/accounts/register/"
 1511381696494	addons.xpi	WARN	Exception running bootstrap method shutdown on activity-stream@mozilla.org: [Exception... "Component returned failure code: 0x80004005 (NS_ERROR_FAILURE) [nsIObserverService.removeObserver]"  nsresult: "0x80004005 (NS_ERROR_FAILURE)"  location: "JS frame :: resource://activity-stream/lib/SnippetsFeed.jsm :: uninit :: line 123"  data: no] Stack trace: uninit()@resource://activity-stream/lib/SnippetsFeed.jsm:123 < onAction()@resource://activity-stream/lib/SnippetsFeed.jsm:138 < _middleware/</<()@resource://activity-stream/lib/Store.jsm:52 < Store/this[method]()@resource://activity-stream/lib/Store.jsm:31 < uninit()@resource://activity-stream/lib/Store.jsm:151 < uninit()@resource://activity-stream/lib/ActivityStream.jsm:266 < uninit()@resource://gre/modules/addons/XPIProvider.jsm -> jar:file:///Applications/Firefox.app/Contents/Resources/browser/features/activity-stream@mozilla.org.xpi!/bootstrap.js:80 < shutdown()@resource://gre/modules/addons/XPIProvider.jsm -> jar:file:///Applications/Firefox.app/Contents/Resources/browser/features/activity-stream@mozilla.org.xpi!/bootstrap.js:196 < callBootstrapMethod()@resource://gre/modules/addons/XPIProvider.jsm:4509 < observe()@resource://gre/modules/addons/XPIProvider.jsm:2281 < GeckoDriver.prototype.quit()@driver.js:3307 < despatch()@server.js:553 < execute()@server.js:527 < onPacket/<()@server.js:502 < onPacket()@server.js:501 < _onJSONObjectReady/<()@transport.js:501
+1511430667540	geckodriver	INFO	geckodriver 0.19.1
+1511430667553	geckodriver	INFO	Listening on 127.0.0.1:56863
+1511430668683	mozrunner::runner	INFO	Running command: "/Applications/Firefox.app/Contents/MacOS/firefox-bin" "-marionette" "-profile" "/var/folders/px/wzvy5s510qgcq285p8dr_w1h0000gn/T/rust_mozprofile.KnmCYocNaY4C"
+1511430669742	Marionette	INFO	Enabled via --marionette
+2017-11-23 09:51:11.471 plugin-container[12327:3508861] *** CFMessagePort: bootstrap_register(): failed 1100 (0x44c) 'Permission denied', port = 0xa33b, name = 'com.apple.tsm.portname'
+See /usr/include/servers/bootstrap_defs.h for the error codes.
+1511430672049	Marionette	INFO	Listening on port 56886
+1511430672077	Marionette	WARN	TLS certificate errors will be ignored for this session
+1511430672174	Marionette	DEBUG	Register listener.js for window 2147483649
+1511430672314	Marionette	DEBUG	Received DOM event "beforeunload" for "about:blank"
+1511430672381	Marionette	DEBUG	Received DOM event "pagehide" for "about:blank"
+1511430672384	Marionette	DEBUG	Received DOM event "unload" for "about:blank"
+1511430672972	Marionette	DEBUG	Received DOM event "DOMContentLoaded" for "http://localhost:56862/accounts/login/"
+1511430673025	Marionette	DEBUG	Received DOM event "pageshow" for "http://localhost:56862/accounts/login/"
+1511430674556	geckodriver	INFO	geckodriver 0.19.1
+1511430674565	geckodriver	INFO	Listening on 127.0.0.1:57065
+1511430676776	mozrunner::runner	INFO	Running command: "/Applications/Firefox.app/Contents/MacOS/firefox-bin" "-marionette" "-profile" "/var/folders/px/wzvy5s510qgcq285p8dr_w1h0000gn/T/rust_mozprofile.C26Kxrio5RV1"
+1511430677476	Marionette	INFO	Enabled via --marionette
+2017-11-23 09:51:18.869 plugin-container[12333:3509265] *** CFMessagePort: bootstrap_register(): failed 1100 (0x44c) 'Permission denied', port = 0xa23b, name = 'com.apple.tsm.portname'
+See /usr/include/servers/bootstrap_defs.h for the error codes.
+1511430679362	Marionette	INFO	Listening on port 57098
+1511430679474	Marionette	WARN	TLS certificate errors will be ignored for this session
+1511430679556	Marionette	DEBUG	Register listener.js for window 2147483649
+1511430679624	Marionette	DEBUG	Received DOM event "beforeunload" for "about:blank"
+1511430679696	Marionette	DEBUG	Received DOM event "pagehide" for "about:blank"
+1511430679697	Marionette	DEBUG	Received DOM event "unload" for "about:blank"
+1511430680151	Marionette	DEBUG	Received DOM event "DOMContentLoaded" for "http://localhost:57064/accounts/login/?next=/"
+1511430680484	Marionette	DEBUG	Received DOM event "pageshow" for "http://localhost:57064/accounts/login/?next=/"
+1511430681162	geckodriver	INFO	geckodriver 0.19.1
+1511430681168	geckodriver	INFO	Listening on 127.0.0.1:57231
+1511430682212	mozrunner::runner	INFO	Running command: "/Applications/Firefox.app/Contents/MacOS/firefox-bin" "-marionette" "-profile" "/var/folders/px/wzvy5s510qgcq285p8dr_w1h0000gn/T/rust_mozprofile.ZCR0wZF9xY1B"
+1511430682956	Marionette	INFO	Enabled via --marionette
+2017-11-23 09:51:24.288 plugin-container[12337:3509769] *** CFMessagePort: bootstrap_register(): failed 1100 (0x44c) 'Permission denied', port = 0xa13f, name = 'com.apple.tsm.portname'
+See /usr/include/servers/bootstrap_defs.h for the error codes.
+1511430684893	Marionette	INFO	Listening on port 57252
+1511430684971	Marionette	WARN	TLS certificate errors will be ignored for this session
+1511430685059	Marionette	DEBUG	Register listener.js for window 2147483649
+1511430685123	Marionette	DEBUG	Received DOM event "beforeunload" for "about:blank"
+1511430685177	Marionette	DEBUG	Received DOM event "pagehide" for "about:blank"
+1511430685177	Marionette	DEBUG	Received DOM event "unload" for "about:blank"
+1511430685627	Marionette	DEBUG	Received DOM event "DOMContentLoaded" for "http://localhost:57230/accounts/register/"
+1511430685820	Marionette	DEBUG	Received DOM event "pageshow" for "http://localhost:57230/accounts/register/"
